pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "watson-ctf"
version = "0.1.0"
description = "A forensics CTF solver CLI — Watson does the legwork so you can make the deductions."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "python-magic>=0.4",
    "Pillow>=10",
    "pyyaml>=6",
    "click>=8",
]

[project.optional-dependencies]
full = [
    "mutagen",
    "pypdf",
    "scapy",
]

[project.scripts]
watson = "watson.cli:main_entry"

[tool.setuptools.packages.find]
where = ["."]
include = ["watson*"]
